Refactor: Fix formating tests
