+++ src/helpers/tests/flow-chart.testData.ts
@@ -49,71 +49,85 @@ export const formattedFlow = {
   elements: [
     {
       id: "gateway",
-      type: "gateway",
+      type: "input",
       position: { y: 150, x: 250 },
       data: {
         label: "gateway",
+        needs: [],
+        send_to: {},
+        properties: {},
         depth: 0,
       },
     },
     {
       id: "segmenter",
-      type: "pod",
+      type: "default",
       position: { y: 300, x: 250 },
       data: {
         label: "segmenter",
         needs: ["gateway"],
-        uses: "pods/segment.yml",
-        read_only: true,
+        send_to: {},
+        properties: {
+          uses: "pods/segment.yml",
+          read_only: true,
+        },
         depth: 1,
       },
     },
     {
       id: "encoder",
-      type: "pod",
+      type: "default",
       position: { y: 450, x: 250 },
       data: {
         label: "encoder",
         needs: ["segmenter"],
-        uses: "pods/encode.yml",
-        polling: "any",
-        timeout_ready: 600000,
-        read_only: true,
+        send_to: {},
+        properties: {
+          uses: "pods/encode.yml",
+          polling: "any",
+          timeout_ready: 600000,
+          read_only: true,
+        },
         depth: 2,
       },
     },
     {
       id: "chunk_idx",
-      type: "pod",
+      type: "default",
       position: { y: 600, x: 250 },
       data: {
         label: "chunk_idx",
         needs: ["encoder"],
-        polling: "any",
-        uses: "pods/chunk.yml",
+        send_to: {},
+        properties: {
+          polling: "any",
+
+          uses: "pods/chunk.yml",
+        },
         depth: 3,
       },
     },
     {
       id: "doc_idx",
-      type: "pod",
+      type: "default",
       position: { y: 300, x: 500 },
       data: {
         label: "doc_idx",
         needs: ["gateway"],
-        polling: "any",
-        uses: "pods/doc.yml",
+        send_to: {},
+        properties: { polling: "any", uses: "pods/doc.yml" },
         depth: 1,
       },
     },
     {
       id: "join_all",
-      type: "pod",
+      type: "default",
       position: { y: 750, x: 250 },
       data: {
         label: "join_all",
         needs: ["doc_idx", "chunk_idx"],
-        uses: "_merge",
+        send_to: {},
+        properties: { uses: "_merge" },
         depth: 4,
       },
     },
@@ -121,37 +135,31 @@ export const formattedFlow = {
       id: "e-gateway-to-segmenter",
       source: "gateway",
       target: "segmenter",
-      type: "step",
     },
     {
       id: "e-segmenter-to-encoder",
       source: "segmenter",
       target: "encoder",
-      type: "step",
     },
     {
       id: "e-encoder-to-chunk_idx",
       source: "encoder",
       target: "chunk_idx",
-      type: "step",
     },
     {
       id: "e-gateway-to-doc_idx",
       source: "gateway",
       target: "doc_idx",
-      type: "step",
     },
     {
       id: "e-doc_idx-to-join_all",
       source: "doc_idx",
       target: "join_all",
-      type: "step",
     },
     {
       id: "e-chunk_idx-to-join_all",
       source: "chunk_idx",
       target: "join_all",
-      type: "step",
     },
   ],
 }
@@ -180,28 +188,28 @@ with:
         y: 750
 pods:
   segmenter:
-    needs: gateway
     uses: pods/segment.yml
     read_only: true
+    needs: gateway
   encoder:
-    needs: segmenter
     uses: pods/encode.yml
     polling: any
     timeout_ready: 600000
     read_only: true
+    needs: segmenter
   chunk_idx:
-    needs: encoder
     polling: any
     uses: pods/chunk.yml
+    needs: encoder
   doc_idx:
-    needs: gateway
     polling: any
     uses: pods/doc.yml
+    needs: gateway
   join_all:
+    uses: _merge
     needs:
       - doc_idx
       - chunk_idx
-    uses: _merge
 `
 
 export const flowArguments: FlowArguments = {
